More work on endpoints

+++ accord-app/src/main/java/com/group11/accord/app/services/AuthorizationService.java
@@ -19,7 +19,7 @@ public class AuthorizationService {
         return "";
     }
 
-    public void changePassword(int accountId, String token, String newPassword, String oldPassword) {
+    public void changePassword(Long accountId, String token, String newPassword, String oldPassword) {
 
     }
 }
